Adds broadcast updated tracker file_table to all peers in peers_list. Updates newpeerip in struct file_table to a string array.
Updates mechanism to append new_peers to newpeerip.
Creates basic broadcast function to send tracker file_table to all peers on tracker file_table update.

packet of type FILE_UPDATE to be used by peer for both cases when monitored directory updated local or when sync completed with another peer.
i.e peer should sends packet of type FILE_UPDATE once it successfully syncs a file updated on another peer.
If the timestamp of the file on tracker is same as that in the peers_file and peer doesn't exist in newpeerip than adds it to files newpeerip.